# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog API",
        "description": "REST API for the Blog application. Provides endpoints for managing posts, categories, tags, comments, and user authentication.",
        "contact": {
            "name": "DigiK Team",
            "email": "admin@digikteam.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{protocol}://{host}",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/tags/{slug}": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Get a single tag by slug",
                "operationId": "d7b588448d1c27b207743c23b8b9dc0f",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Tag slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tag not found"
                    }
                }
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {your_token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}